Change type to avoid integer promotion confusion.
The compiler promotes `uint8_t` to `int` when performing bitwise
operations, which then gets confusing when you assign to a `size_t`
variable.  Avoid the whole mess by using `unsigned` instead of
`uint8_t`.

+++ core/Lucy/Analysis/StandardTokenizer.c
@@ -278,7 +278,7 @@ word_break:
 
 static int
 S_wb_lookup(const char *ptr) {
-    uint8_t start = *(uint8_t*)ptr++;
+    unsigned start = *(uint8_t*)ptr++;
 
     if (start < 0x80) { return wb_ascii[start]; }
